chore: rename ShiftFormInput to ShiftFormNote
